backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
chromadb
requests
litellm
gTTS
Pillow
sqlalchemy
psycopg2-binary
redis
stripe
facebook-sdk
ShopifyAPI
apscheduler
pytest
pytest-cov
httpx
Faker
python-jose[cryptography]
passlib[bcrypt]
